Create WarpNetwork
